+++ src/kagglehub/resolver.py
@@ -13,7 +13,7 @@ class Resolver(Generic[T]):
     __metaclass__ = abc.ABCMeta
 
     def __call__(
-        self, handle: T, path: Optional[str] = None, *, force_download: Optional[bool] = False
+        self, handle: T, path: Optional[str] = None, *, force_download: Optional[bool] = False, **kwargs
     ) -> tuple[str, Optional[int]]:
         """Resolves a handle into a path with the requested file(s) and the resource's version number.
 
@@ -26,7 +26,7 @@ class Resolver(Generic[T]):
             A tuple of: (string representing the path, version number of resolved datasource if present)
             Some cases where version number might be missing: Competition datasource, API-based models.
         """
-        path, version = self._resolve(handle, path, force_download=force_download)
+        path, version = self._resolve(handle, path, force_download=force_download, **kwargs)
 
         # Note handles are immutable, so _resolve() could not have altered our reference
         register_datasource_access(handle, version)
@@ -35,7 +35,7 @@ class Resolver(Generic[T]):
 
     @abc.abstractmethod
     def _resolve(
-        self, handle: T, path: Optional[str] = None, *, force_download: Optional[bool] = False
+        self, handle: T, path: Optional[str] = None, *, force_download: Optional[bool] = False, **kwargs
     ) -> tuple[str, Optional[int]]:
         """Resolves a handle into a path with the requested file(s) and the resource's version number.